refactor index.ts naming to auth-controller.ts for better readability
