+++ backend/user/src/routes/auth.ts
@@ -1,8 +1,8 @@
 import express from 'express';
 
-import { login, logout } from '@/controllers/auth';
+import { login, logout } from '@/controllers/auth/auth-controller';
 import { limiter } from '@/lib/ratelimit';
-import { register } from '@/controllers/auth';
+import { register } from '@/controllers/auth/auth-controller';
 
 const router = express.Router();
